backend/requirements.txt
# Flask core
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Babel==4.0.0

# Database
psycopg2-binary==2.9.9
pgvector==0.2.4
SQLAlchemy==2.0.23

# Redis & Caching
redis==5.0.1

# Celery for async tasks
celery==5.3.4

# Object storage (S3/MinIO)
minio==7.2.0
boto3==1.34.10

# Authentication & Security
PyJWT==2.8.0
bcrypt==4.1.2
python-dotenv==1.0.0

# Media processing (lightweight)
Pillow==10.1.0

# API utilities
marshmallow==3.20.1
webargs==8.3.0

# NOTE: Heavy ML dependencies removed for development
# Add these back when implementing specific features:
#
# For transcription:
#   openai-whisper==20231117 (requires ~3GB+ torch)
#
# For local embeddings:
#   sentence-transformers==2.2.2 (requires torch)
#   numpy==1.26.2
#
# For OpenAI API (lightweight, recommended):
#   openai==1.6.1
#
# We'll use OpenAI API for embeddings when needed (no local models)

# Validation
email-validator==2.1.0

# Utils
python-slugify==8.0.1
requests==2.31.0
gunicorn==21.2.0

# Markdown rendering
markdown==3.5.1
pymdown-extensions==10.7
